Split snapshot into a package; add FileSnapshot factories
Splits the monolithic snapshot.py into pyfr/snapshot/{base,intg,file,
soln,stats,region,sample}.py. FileSnapshot becomes the abstract base for
file-loaded snaps with two classmethod entry points: from_file(meshf,
solnf) opens .pyfrm + .pyfrs from disk and dispatches on data/prefix;
from_loaded(mesh, soln) wraps already-loaded objects (writers). All
callers updated to the canonical entry points.
